move main.go to cmd/server/
