+++ run_server.nix.sh
@@ -2,4 +2,4 @@
 #! nix-shell -i bash -p go_1_17 -p reflex
 #! nix-shell --pure --quiet
 
-reflex -r "(main.go|(src|data)/)" -s -- go run -race . --reload
+reflex -r "(cmd|src|data)/" -s -- go run -race ./cmd/server --reload
